Extract workspace package import ZIP helpers
